third_party/WebKit/LayoutTests/shadow-dom/crashes/offsetParent-layoutObject-lifecycle.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script>
// This case is created from reproduction case by cloudfuzzer.
window.addEventListener('load', () => {
  test(() => {
    var link = document.createElement('link');
    link.setAttribute('rel','import');
    document.head.appendChild(link);

    var style = document.createElement('style');
    style.textContent = '@import url(x)';
    document.head.appendChild(style);

    var frame = document.createElement('frame');
    document.body.appendChild(frame);

    // This could cause crash.
    frame.offsetHeight;
    assert_true(true, 'This test must run without crash.');
  }, "HTMLElement.offsetHeight should not cause crash.");
}, false);
</script>
<body></body>
